requirements.txt
fastapi
uvicorn[standard]
python-multipart
pypdf
python-docx
pillow
pytesseract
pandas
sentence-transformers
faiss-cpu
requests
pydantic
Streamlit 